extract ClaimButton component
Reduces redundant code and extracts logic/handling for the claim button on any given project. Switches to using `formatAmount` helper in `utils/amounts` for consistency.
